# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Story Generation Workflow API
  version: 1.0.0
  description: |
    Design-first specification for the Story Generation Workflow service. This document captures
    the currently implemented Express route handlers and internal services that orchestrate story
    creation, editing, media generation, audiobook production, and workflow coordination.
servers:
  - url: https://story-generation-workflow.example.com
    description: Production
  - url: http://localhost:8080
    description: Local development
security:
  - {}
...
paths:
...
  /audio/create-audiobook:
    post:
      tags: [Audio]
      summary: Trigger audiobook generation
      description: Starts the asynchronous audiobook generation workflow for the specified story.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AudiobookTriggerRequest'
            examples:
              default:
                value:
                  storyId: f2b0c4e2-05d4-4f74-b0a1-7f6314b8f5f3
                  voice: nova
      responses:
        '200':
          description: Workflow started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudiobookTriggerResponse'
        '400':
          description: storyId missing from payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to enqueue the workflow.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /audio/internal/audiobook/chapter:
    post:
      tags: [Audio]
      summary: Generate chapter audio (external route)
      description: Generates chapter audio and stores the resulting asset metadata. Requires API key even though used by internal automation.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChapterAudioRequest'
      responses:
        '200':
          description: Chapter audio generated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterAudioResponse'
        '400':
          description: Missing required parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Generation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /audio/internal/audiobook/finalize:
    post:
      tags: [Audio]
      summary: Finalize audiobook (external route)
      description: Aggregates generated chapter audio files, updates story metadata, and returns a summary.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AudiobookFinalizeRequest'
      responses:
        '200':
          description: Finalization succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudiobookFinalizeResponse'
        '500':
          description: Finalization failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /internal/audiobook/chapter:
    post:
      tags: [Internal]
      summary: Generate chapter audio (internal)
      description: Generates audio for a chapter without requiring an external API key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChapterAudioRequest'
      responses:
        '200':
          description: Chapter audio generated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterAudioResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Generation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /internal/audiobook/finalize:
    post:
      tags: [Internal]
      summary: Finalize audiobook (internal)
      description: Updates story metadata once all chapters have been narrated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AudiobookFinalizeRequest'
      responses:
        '200':
          description: Audiobook metadata updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudiobookFinalizeResponse'
        '500':
          description: Finalization failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
...
  schemas:
...
    ErrorResponse:
      type: object
      description: Standard error envelope returned across the API. Fields may vary slightly by route.
      properties:
        success:
          type: boolean
          description: Indicates success when present. Typically omitted for internal-only endpoints that predate the convention.
          example: false
        error:
          type: string
          description: Human-readable error description.
        message:
          type: string
          description: Optional detailed error message.
        details:
          type: object
          additionalProperties: true
          description: Optional machine-readable context.
      required:
        - error
...
    AudiobookTriggerRequest:
      type: object
      required:
        - storyId
      properties:
        storyId:
          type: string
          format: uuid
        voice:
          type: string
    AudiobookTriggerResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        storyId:
          type: string
          format: uuid
        voice:
          type: string
        status:
          type: string
        executionId:
          type: string
    ChapterAudioRequest:
      type: object
      required:
        - storyId
        - chapterNumber
        - chapterContent
        - storyTitle
      properties:
        storyId:
          type: string
          format: uuid
        chapterNumber:
          type: integer
          minimum: 1
        chapterTitle:
          type: string
        chapterContent:
          type: string
        storyTitle:
          type: string
        storyAuthor:
          type: string
        dedicatoryMessage:
          type: string
        voice:
          type: string
        storyLanguage:
          type: string
        isFirstChapter:
          oneOf:
            - type: boolean
            - type: string
    ChapterAudioResponse:
      type: object
      properties:
        success:
          type: boolean
        chapterNumber:
          type: integer
        audioUrl:
          type: string
        duration:
          type: number
        format:
          type: string
        provider:
          type: string
        voice:
          type: string
        metadata:
          type: object
          additionalProperties: true
    AudiobookFinalizeRequest:
      type: object
      required:
        - storyId
      properties:
        storyId:
          type: string
          format: uuid
    AudiobookFinalizeResponse:
      type: object
      properties:
        success:
          type: boolean
        storyId:
          type: string
          format: uuid
        audioUrls:
          type: object
          additionalProperties:
            type: string
        totalDuration:
          type: number
        chaptersProcessed:
          type: integer
...